# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SaaS Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: ["string", "null"]
        - name: clientId
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
    patch:
      operationId: updateConversation
      tags: [conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
    delete:
      operationId: deleteConversation
      tags: [conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Clients ────────────────────────────────────────────────
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Attendances ────────────────────────────────────────────
  /attendances:
    get:
      operationId: listAttendances
      tags: [attendances]
      summary: List attendances
      responses:
        "200":
          description: Attendances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendances]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendances/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendances]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
    delete:
      operationId: deleteAttendance
      tags: [attendances]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Agents ─────────────────────────────────────────────────
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Users ──────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Plans ──────────────────────────────────────────────────
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List plans
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    patch:
      operationId: updatePlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Financial ──────────────────────────────────────────────
  /financial:
    get:
      operationId: listTransactions
      tags: [financial]
      summary: List financial transactions
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [financial]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /financial/{id}:
    patch:
      operationId: updateTransaction
      tags: [financial]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ── Integrations ───────────────────────────────────────────
  /integrations:
    get:
      operationId: listIntegrations
      tags: [integrations]
      summary: List integrations
      responses:
        "200":
          description: Integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Integration"
    post:
      operationId: createIntegration
      tags: [integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
  /integrations/{id}:
    patch:
      operationId: updateIntegration
      tags: [integrations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Integration"
    delete:
      operationId: deleteIntegration
      tags: [integrations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Reports ────────────────────────────────────────────────
...
components:
  schemas:
...
    Conversation:
      type: object
      required: [id, clientId, status, channel, lastMessage, createdAt, updatedAt]
      properties:
        id: { type: string }
        clientId: { type: string }
        clientName: { type: ["string", "null"] }
        agentId: { type: ["string", "null"] }
        agentName: { type: ["string", "null"] }
        status: { type: string }
        channel: { type: string }
        lastMessage: { type: string }
        unreadCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    ConversationInput:
      type: object
      required: [clientId, channel]
      properties:
        clientId: { type: string }
        agentId: { type: ["string", "null"] }
        channel: { type: string }
        status: { type: string }
    ConversationUpdate:
      type: object
      properties:
        agentId: { type: ["string", "null"] }
        status: { type: string }

    # ── Client ────────────────────────────────────────────────
    Client:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        status: { type: string }
        totalConversations: { type: integer }
        createdAt: { type: string }
    ClientInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        company: { type: string }
        status: { type: string }
    ClientUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        status: { type: string }

    # ── Attendance ────────────────────────────────────────────
    Attendance:
      type: object
      required: [id, status, createdAt]
      properties:
        id: { type: string }
        conversationId: { type: ["string", "null"] }
        clientId: { type: ["string", "null"] }
        clientName: { type: ["string", "null"] }
        agentId: { type: ["string", "null"] }
        agentName: { type: ["string", "null"] }
        status: { type: string }
        channel: { type: ["string", "null"] }
        startedAt: { type: ["string", "null"] }
        endedAt: { type: ["string", "null"] }
        durationSeconds: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    AttendanceInput:
      type: object
      required: [status]
      properties:
        conversationId: { type: string }
        clientId: { type: string }
        agentId: { type: string }
        channel: { type: string }
        status: { type: string }
        notes: { type: string }
    AttendanceUpdate:
      type: object
      properties:
        status: { type: string }
        agentId: { type: ["string", "null"] }
        endedAt: { type: ["string", "null"] }
        durationSeconds: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }

    # ── Agent ─────────────────────────────────────────────────
...